+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## develop
 
+### Fixes
+
+* `Pipeline.Okay()` now returns `(bool, error)`
+
 ## v1.5.0
 
 Released Tuesday, 17th September 2019.